+++ src/sample/View/SVGTextButtonView.xaml.cs
@@ -1,0 +1,22 @@
+﻿using System.Windows;
+using System.Windows.Controls;
+
+
+namespace SWUI.Sample.View
+{
+    /// <summary>
+    /// Logique d'interaction pour SVGTextButtonView.xaml
+    /// </summary>
+    public partial class SVGTextButtonView : UserControl
+    {
+        public SVGTextButtonView()
+        {
+            InitializeComponent();
+        }
+
+        private void RadioButton_Checked(object sender, RoutedEventArgs e)
+        {
+
+        }
+    }
+}