+++ changelog.md
@@ -1,12 +1,28 @@
-### SWUI 1.1.0
+## SWUI 2.0.0pre
+  <br/>
 
-Update StateButtons:
-- fix BaseColor property update
-- add OverColor property
+**Changes**  
+    SVGButton : Change type of OverBackgroungcolor property. 
+    This allow to use gradient brush.
+  
+**New feature**  
+    New control : SVGTextButton
+
+**Fixes**  
+    StateButtons : icons are not always updated #2
+<br/><br/><br/>
+
+## SWUI 1.1.0
+
+**Fixes**  
+    StateButtons : fix BaseColor property update
+
+**New feature**  
+    StateButtons : add OverColor property
 <br/><br/><br/>
   
   
-### SWUI 1.0.0
+## SWUI 1.0.0
 
 First release with :
 - StateButtons